requirements.txt
# Requirements

Pillow==5.3.0
